+[core] Type-safe Message Passing
After much wondering and going down the GADT/Heterogeneous List pathI
found accidentally extensible variants, which allowed me to model the
messages in a way that a homogenous list could be used for the
mailboxes, yet keeping each data constructor internal to each process
definition.

On top of that, TestModule_3 exhibits 2 long-lived processes
collaborating to achieve a goal: print the current timestamp.

The first one is a Logger process, that responds to `Log(int)` messages
and logs to the console. The second one is a Clock process, that is
configurable and will send a clock tick periodically to the configured
address (pid).

This makes for a proof of concept. The next step is to model more
complex behavior (like a small game) and port some OTP idioms for
reliability.