[core] Strip out scheduler
In this iteration I'm taking an approach where a Coordinator will take
care of spawning separate Workers and communicating with them.

A Policy, passe in to the Scheduler, will define different properties
that the Coordinator (and possibly the Workers in the future) will have;
such as how to balance work (stealing vs sharing), or how many workers
to utilize (full SMP, or limited number of cores).

A Worker registry makes it trivial to find a worker by it's id and gain
access to the low-level communication implemented on top of the Unix module.

Still a work in progress.