Type-safe message passing across worker boundaries
This attempt at getting type-safe message passing across worker
boundaries had a few concerns involved.

Unfortunately, the `Message.t` extensible variant being used before
loses type information across process boundaries due to a design
decision in the `Marshal` module. This meant that while messages
type-checked as valid to send, and arrived perfectly at the workers,
they had lost the information that made pattern-matching on them
possible and thus were unconsumable.

Removing this variant meant parametrizing `Pid.t` and `Process.t` with
the type of messages that will be received by the process, so that `<-`
could type-check at the call site that we would only send a message to a
process that will actually handle that message.

This meant refactoring the `Registry.re` module, that is, the Process
Registry, to become a heterogeneous collection that can keep several
`Procees.t('m)` together, where each `'m` may be a different message
time.

The refactoring process tried out:

1. `CCHet.Tbl`,
2. `Hmap`, and
3. `Gmap`

All of which have their draw backs. The first 2 were able to encode
uniqueness of all types and thus allow reactor to have type-safe message
passing _on a per process basis_, and the 3rd one relied on a bounded
type-universe of messages known upfront.

Because reactor as a library does not know in advance the types of all
the messages that are being sent in it, the 3rd option was originally
discarded.

Options 1 and 2 have worked perfectly fine in the context of a single
process, but since they rely on first-class modules and extensible
variants, they suffer the same problems that the original solution.