[scheduler] Worker should enqueue tasks normally
Before this changeset, refering to the `Scheduler.spawn` and
`Scheduler.send` meant that the process would act as if it was a
_coordinator_.

That is not the case, specially when trying to send messages and spawn
actors from a Worker unix process.

With this change, calling `Scheduler.spawn` or `Scheduler.send` will
first check if it its in a coordinator or a worker, and proceed to queue
up the action accordingly.